pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=0.3.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
    "ty>=0.0.16",
    "transformers>=4.52.0",
    "torch>=2.6.0",
    "peft>=0.15.2",
    "torchvision",
    "pillow>=10.0.0",
    "langchain-openai>=1.1.9",
    "pypdf>=6.7.0",
    "jina>=3.34.0",
    "openai>=2.20.0",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "huggingface-hub>=0.36.2",
    "langchain-core>=1.2.11",
    "langchain-groq>=1.1.2",
]
